Handle github_token responses properly
When a gh token is passed with `--gitub-token`, we need to send it to
the API (`/auth/github`) before we know whether or not it has sufficient
scopes.

If the token has correct scopes, API returns `access_token` we can use.
If the token does not have sufficient scopes, at the moment, it is a
redirect, which, when followed, is a success (200) but with unusable
data.
As `#post_raw` returns a String in such a case (`JSON.parse` fails and
body is returned), so we'll just ensure that the response was what we
expect.